Moved attribute vuex module to product module
