+++ core/modules/product/helpers/optionLabel.ts
