+++ core/modules/product/store/attribute/getters.ts
@@ -1,6 +1,6 @@
 import { GetterTree } from 'vuex'
-import AttributeState from './types/AttributeState'
-import RootState from '../../types/RootState'
+import AttributeState from '../../types/AttributeState'
+import RootState from '@vue-storefront/store/types/RootState'
 
 const getters: GetterTree<AttributeState, RootState> = {
   attributeListByCode: (state) => state.list_by_code,
