+++ core/modules/product/store/attribute/mutation-types.ts
@@ -1,0 +1,2 @@
+export const SN_ATTRIBUTE = 'attribute'
+export const ATTRIBUTE_UPD_ATTRIBUTES = SN_ATTRIBUTE + '/UPD_ATTRIBUTES'
