+++ core/modules/product/types/AttributeState.ts
